Code review 1 changes
