+++ internal/cmd/controller/reconciler/bundle_controller_test.go
@@ -33,7 +33,6 @@ import (
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
-	"sigs.k8s.io/controller-runtime/pkg/event"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 )
 
@@ -1192,72 +1191,6 @@ func TestBundleDeploymentMapFunc(t *testing.T) {
 		reqs := mapFunc(context.Background(), bd)
 		if diff := cmp.Diff(expected, reqs); diff != "" {
 			t.Errorf("mismatch (-want +got):\n%s", diff)
-		}
-	})
-}
-
-func TestBundleDeploymentStatusChangedPredicate(t *testing.T) {
-	p := reconciler.BundleDeploymentStatusChangedPredicate()
-
-	t.Run("Create", func(t *testing.T) {
-		e := event.CreateEvent{Object: &fleetv1.BundleDeployment{}}
-		if !p.Create(e) {
-			t.Error("expected true for create event")
-		}
-	})
-
-	t.Run("Delete", func(t *testing.T) {
-		// Delete is not defined, so it should be false
-		e := event.DeleteEvent{Object: &fleetv1.BundleDeployment{}}
-		if !p.Delete(e) {
-			t.Error("expected true for delete event")
-		}
-	})
-
-	t.Run("Generic", func(t *testing.T) {
-		// Generic is not defined, so it should be false
-		e := event.GenericEvent{Object: &fleetv1.BundleDeployment{}}
-		if p.Generic(e) {
-			t.Error("expected false for generic event")
-		}
-	})
-
-	t.Run("Update", func(t *testing.T) {
-		oldBD := &fleetv1.BundleDeployment{
-			Status: fleetv1.BundleDeploymentStatus{Ready: false},
-		}
-		newBD := oldBD.DeepCopy()
-
-		// No change
-		e := event.UpdateEvent{ObjectOld: oldBD, ObjectNew: newBD}
-		if p.Update(e) {
-			t.Error("should be false when status is identical")
-		}
-
-		// Status changed
-		newBD.Status.Ready = true
-		e = event.UpdateEvent{ObjectOld: oldBD, ObjectNew: newBD}
-		if !p.Update(e) {
-			t.Error("should be true when status changes")
-		}
-
-		// Deletion timestamp added
-		newBD = oldBD.DeepCopy()
-		now := metav1.Now()
-		newBD.DeletionTimestamp = &now
-		e = event.UpdateEvent{ObjectOld: oldBD, ObjectNew: newBD}
-		if !p.Update(e) {
-			t.Error("should be true when deletion timestamp is set")
-		}
-
-		// Nil objects
-		e = event.UpdateEvent{ObjectOld: nil, ObjectNew: newBD}
-		if p.Update(e) {
-			t.Error("should be false for nil old object")
-		}
-		e = event.UpdateEvent{ObjectOld: oldBD, ObjectNew: nil}
-		if p.Update(e) {
-			t.Error("should be false for nil new object")
 		}
 	})
 }
